Backend/AMY-Plotter-Interview-App-master/requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-xml==2.0.0
importlib-metadata==3.1.1
Markdown==3.3.3
pytz==2020.4
sqlparse==0.4.1
zipp==3.4.0
